Reintroduced generic type in Expression class
git-svn-id: https://svn.code.sf.net/p/benerator/code/benerator/trunk@3063 37d91da1-6759-5a44-b2bd-541c34f01409

+++ src/test/java/org/databene/benerator/engine/parser/ParseRunTaskTest.java
@@ -52,8 +52,8 @@ public class ParseRunTaskTest {
         RunTaskParser parser = new RunTaskParser();
 		RunTaskStatement task = parser.parse(doc.getDocumentElement(), new ResourceManagerSupport());
 		BeneratorContext context = new BeneratorContext();
-		assertEquals(5, task.getCount().evaluate(context));
-		assertEquals(2, task.getPageSize().evaluate(context));
+		assertEquals(5L, task.getCount().evaluate(context).longValue());
+		assertEquals(2L, task.getPageSize().evaluate(context).longValue());
 		assertEquals(new PageListenerMock(1), task.getPager().evaluate(context));
 		task.execute(new BeneratorContext());
 		assertEquals(5, TaskMock.count.get());
